In an effort to support 32-bit builds (I only care about this because of gdebugger, which may not work anyway), I cleaned up a warning

+++ Source/PowerProps.cpp
@@ -733,7 +733,7 @@ public:
 
 			case PT_ENUM:
 				SetEnumStrings(pprop->GetEnumStrings());
-				SetEnumVal(pprop->AsInt());
+				SetEnumVal((size_t)(pprop->AsInt()));
 				break;
 		}
